Correct tree type issues
