+++ packages/components/src/Tree/index.ts
@@ -28,3 +28,4 @@ export * from './Tree'
 export * from './TreeArtificial'
 export * from './TreeItem'
 export * from './TreeGroup'
+export * from './types'
